deployment02/backend/requirements.txt
langchain
datasets==2.13.0
peft==0.4.0
bitsandbytes==0.40.2
trl==0.4.7
safetensors>=0.3.1

faiss-cpu
tiktoken
sentence-transformers
einops
huggingface-hub
xformers
auto-gptq
optimum
accelerate

Flask==2.2.5
flask-cors
# flask-ngrok
# pyngrok

gunicorn==20.1.0
